Update PrivateRoute & make Layout respond to AuthStatus
