+++ src/App/App.tsx
@@ -1,30 +1,24 @@
 import MainPage from '../pages/Main/Main';
-import { Addresses, AuthorizationStatus } from '../const';
+import { Addresses } from '../const';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Login } from '../pages/Login/Login';
 import { Favorites } from '../pages/Favorites/Favorites';
 import { Offer } from '../pages/Offer/Offer';
 import { NotFound } from '../pages/404/404';
-import PrivateRoute from '../components/PrivateRoute/PrivateRoute';
+import { PrivateRoute } from '../components/PrivateRoute/PrivateRoute';
 import { MainContent } from '../pages/Main/MainContent';
-import { AuthStatus, State } from '../types';
-import { useSelector } from 'react-redux';
 
-export const App = () => {
-  const authStatus: AuthStatus = useSelector((state: State) => state.authorizationStatus);
+export const App = () => (
+  <BrowserRouter>
+    <Routes>
+      <Route path={Addresses.Main} element={<MainPage/>}>
+        <Route path=':city' element={<MainContent/>}></Route>
+      </Route>
+      <Route path={Addresses.Login} element={<Login/>} />
+      <Route path={Addresses.Favorites} element={<PrivateRoute><Favorites/></PrivateRoute>} />
+      <Route path={Addresses.Offer} element={<Offer/>} />
+      <Route path={'*'} element={<NotFound/>} />
+    </Routes>
+  </BrowserRouter>
 
-  return(
-    <BrowserRouter>
-      <Routes>
-        <Route path={Addresses.Main} element={<MainPage/>}>
-          <Route path=':city' element={<MainContent/>}></Route>
-        </Route>
-        <Route path={Addresses.Login} element={<Login/>} />
-        <Route path={Addresses.Favorites} element={<PrivateRoute authorizationStatus={authStatus}><Favorites/></PrivateRoute>} />
-        <Route path={Addresses.Offer} element={<Offer/>} />
-        <Route path={'*'} element={<NotFound/>} />
-      </Routes>
-    </BrowserRouter>
-
-  );
-};
+);
