+++ src/components/Layout/AuthUser.tsx
@@ -1,0 +1,24 @@
+import { Link } from 'react-router-dom';
+import { Addresses } from '../../const';
+
+export const AuthUser = () => (
+  <ul className="header__nav-list">
+    <li className="header__nav-item user">
+      <Link
+        className="header__nav-link header__nav-link--profile"
+        to={Addresses.Favorites}
+      >
+        <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+        <span className="header__user-name user__name">
+                  Oliver.conner@gmail.com
+        </span>
+        <span className="header__favorite-count">3</span>
+      </Link>
+    </li>
+    <li className="header__nav-item">
+      <Link className="header__nav-link" to={Addresses.Login}>
+        <span className="header__signout">Sign out</span>
+      </Link>
+    </li>
+  </ul>
+);
