+++ src/components/Layout/Layout.tsx
@@ -1,53 +1,56 @@
 import { Link } from 'react-router-dom';
 import { Addresses } from '../../const';
+import { useEffect } from 'react';
+import { loadAuthStatus } from '../../store/action';
+import { AnyAction, ThunkDispatch } from '@reduxjs/toolkit';
+import { AuthStatus, State } from '../../types';
+import { useDispatch, useSelector } from 'react-redux';
+import { AuthUser } from './AuthUser';
+import { NoAuthUser } from './NoAuthUser';
 
 type LayoutProps = {
   children: JSX.Element;
 }
 
-export const Layout = ({children}: LayoutProps) => (
-  <div className='page page--gray'>
-    <header className="header">
-      <div className="container">
-        <div className="header__wrapper">
-          <div className="header__left">
-            <Link to={Addresses.Main} className="header__logo-link header__logo-link--active">
-              <img
-                className="header__logo"
-                src="img/logo.svg"
-                alt="6 cities logo"
-                width={81}
-                height={41}
-              />
-            </Link>
-          </div>
-          <nav className="header__nav">
-            <ul className="header__nav-list">
-              <li className="header__nav-item user">
-                <Link
-                  className="header__nav-link header__nav-link--profile"
-                  to={Addresses.Favorites}
-                >
-                  <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                  <span className="header__user-name user__name">
-                  Oliver.conner@gmail.com
-                  </span>
-                  <span className="header__favorite-count">3</span>
-                </Link>
-              </li>
-              <li className="header__nav-item">
-                <Link className="header__nav-link" to={Addresses.Login}>
-                  <span className="header__signout">Sign out</span>
-                </Link>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </div>
-    </header>
-    <main>
-      {children}
-    </main>
-  </div>
+export const Layout = ({ children }: LayoutProps) => {
+  const dispatch: ThunkDispatch<State, void, AnyAction> = useDispatch();
+  const authStatus: AuthStatus = useSelector((state: State) => state.authorizationStatus);
+  const isAuth = authStatus === 'AUTH';
 
-);
+  useEffect(() => {
+    if (authStatus === 'UNKNOWN') {
+      dispatch(loadAuthStatus());
+    }
+  }, [authStatus, dispatch]);
+
+  return (
+    <div className='page page--gray'>
+      <header className="header">
+        <div className="container">
+          <div className="header__wrapper">
+            <div className="header__left">
+              <Link to={Addresses.Main} className="header__logo-link header__logo-link--active">
+                <img
+                  className="header__logo"
+                  src="img/logo.svg"
+                  alt="6 cities logo"
+                  width={81}
+                  height={41}
+                />
+              </Link>
+            </div>
+            <nav className="header__nav">
+              {isAuth ? <AuthUser /> : <NoAuthUser />}
+            </nav>
+          </div>
+        </div>
+      </header>
+      <main>
+        {children}
+      </main>
+    </div>
+
+  );
+};
+
+
